Add files via upload
To enable hghlighting 2 sets of SNPs (for  example, based on their novelty or p value) with 2 different colors. Adds a new function "manhattan_2" to be used together with the qqman R package. It is a modified version of the qqman "manhattan" function (https://github.com/stephenturner/qqman/blob/master/R/manhattan.R)